ok for building the imgs
