+++ SRC/radiolabdocker_config.json
@@ -12,7 +12,7 @@
             "BASE" : "ubuntu:20.04"
         },
         "commands" : [
-            "sed -i -e 's%http://archive.ubuntu.com/ubuntu%mirror://mirrors.ubuntu.com/mirrors.txt%' -e 's/^deb-src/#deb-src/' /etc/apt/sources.list",
+            "sed -i -e 's/archive.ubuntu.com/mirrors.ustc.edu.cn/g' -e 's/security.ubuntu.com/mirrors.ustc.edu.cn/g' -e 's/^deb-src/#deb-src/' /etc/apt/sources.list",
             "apt-get update -qq",
             "apt-get install -y",
             "*    wget gnupg xvfb x11-xserver-utils python3-pyinotify mesa-utils",
@@ -451,7 +451,7 @@
         ]
     },
 
-    "all" : {
+    "docker" : {
         "img" : {
             "base" : "radiolab_base",
             "tag" : "latest"
@@ -553,8 +553,8 @@
         ],
         "copy_file" : [
             {
-            "src" : "SRC/bashrc",
-            "dist" : "/opt/bash_config/bashrc"
+            "src" : "SRC/bash_config",
+            "dist" : "/opt/bash_config"
             }
         ]
     }
