--- radiolabdocker/CreateContainer.py
@@ -1,1 +1,0 @@
-import docker
